QLocale: Fixed double to currency string conversion on Mac.
This also fixes the autotest failure on mac with macDefaultLocale test.

+++ src/corelib/tools/qlocale_mac.mm
@@ -318,7 +318,7 @@ static QString macFormatCurrency(const QSystemLocale::CurrencyToStringArgument &
         break;
     }
     case QVariant::Double: {
-        double v = arg.value.toInt();
+        double v = arg.value.toDouble();
         value = CFNumberCreate(NULL, kCFNumberDoubleType, &v);
         break;
     }
